Fix material reordering for NetGen meshes.
There is legacy code that reorders the material indices,
placing the largest material (by element count) at the first
index. For those who know which material is which based on
the NetGen .geo file, this reordering breaks the model since
it can not be easily undone. This change saves a version of
the material indices in fwd_model unmodified. For backwards
compatibility, the reordering remains as the function's
second output variable.

In general, this material reordering aught to be removed in
some future commit since it can break if the region is
broken into eqyal parts or a region with very high element
density exists.

git-svn-id: svn+ssh://svn.code.sf.net/p/eidors3d/code/trunk@4339 82006732-fb7e-4a22-9f28-86a212cce563